requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pymupdf>=1.23.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
plotly>=5.20.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
numpy>=1.26.0
